+++ mtpy/modeling/ModEM/__init__.py
@@ -2,5 +2,6 @@ from .exception import ModEMError, DataError
 from .station import Stations
 from .data import Data
 from .model import Model
+from .residual import Residual
 
-__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model']
+__all__ = ['ModEMError', 'DataError', 'Stations', 'Data', 'Model', 'Residual']
