+++ mtpy/modeling/ModEM/model.py
@@ -1,4 +1,3 @@
-# !/usr/bin/env python
 """
 ==================
 ModEM
